printf(%ul) behaves badly with 32bit vs. 64-bit compilation
%ul is uint64_t in 64-bit environments, but in 32-bit environments, it's
not uint64_t. So this was a problem when compiling for 32-bit
architecutes. We now force conversion to (unsigned long int) of the
operand.

+++ Solver/Solver.cpp
@@ -1563,7 +1563,7 @@ void Solver::dumpSortedLearnts(const char* file, const uint32_t maxSize)
         #endif //STATS_NEEDED
     }
 
-    fprintf(outfile, "c conflicts %lu\n", conflicts);
+    fprintf(outfile, "c conflicts %lu\n", (unsigned long)conflicts);
 
     fprintf(outfile, "c \nc ---------------------------------\n");
     fprintf(outfile, "c learnt clauses from binaryClauses\n");
